Back-End/requirements.txt
djangorestframework
django
mysqlclient
django-cors-headers
